// HomeHub.Web/src/app/core/models/announcements/main-announcement.model.ts
import { Currency } from '../currency.model';
import { AdditionalDetail } from '../lookups/additional-detail.model';
import { AnnouncementPhoto } from './announcement-form.model';

export type MainAnnouncement = {
    id: number;
    subjectTypeId: number;
    subjectTypeName: string;
    typeId: number;
    typeName: string;
    title: string;
    description: string;
    priceAmount: number;
    priceCurrency: Currency;
    advertiserTypeId: number;
    advertiserTypeName: string;
    marketTypeId: number;
    marketTypeName: string;
    area: number;
    numberOfRooms: number;
    buildingFinishConditionId: number;
    buildingFinishConditionName: string;
    buildingMaterialId: number;
    buildingMaterialName: string;
    developmentTypeId: number;
    developmentTypeName: string;
    floorId: number;
    floorName: string;
    numberOfFloors: number;
    heatingTypeId: number;
    heatingTypeName: string;
    buildYear: number;
    ownershipFormId: number;
    ownershipFormName: string;
    windowTypeId: number;
    windowTypeName: string;
    availableFrom: Date;
    advertiserName: string;
    advertiserEmail: string;
    advertiserPhoneNumber: string;
    createdBy: string;
    address: string;
    latitude: number;
    longitude: number;
    videoUrl: string;
    virtualWalkUrl: string;
    photos: AnnouncementPhoto[];
    additionalDetails: AdditionalDetail[];
    isFavorite: boolean;
};